Move UnitUpgradeMenu to popups package
