+++ core/src/com/unciv/ui/screens/overviewscreen/UnitOverviewTab.kt
@@ -32,6 +32,7 @@ import com.unciv.ui.components.extensions.toPrettyString
 import com.unciv.ui.components.input.onClick
 import com.unciv.ui.images.IconTextButton
 import com.unciv.ui.images.ImageGetter
+import com.unciv.ui.popups.UnitUpgradeMenu
 import com.unciv.ui.screens.basescreen.BaseScreen
 import com.unciv.ui.screens.pickerscreens.PromotionPickerScreen
 import com.unciv.ui.screens.pickerscreens.UnitRenamePopup
